Mummi and bilbo contribution!! Curitiba contribution!! veriverse contribution!! LCTV contribution!! catherinahelena (catherinacortega), sjpablo (Pawel Kulesza), fuel (Benjamin Kuhl), r00tkiller (Alexandre),  vapaamies (Vladislav Javadov), epryka (Vova - Vladimir), theaifam5 (Mateus), doraemon (currently... also gosu - hiiamgosu,  dum - noticemesenpaiiamdum..., litAsHellhahah, icarryyourcurry, scorpioinyour...), phanxgames (Henry Price), mushmouth (Patrick McFarling),  sonnystar (sonnystaminecraft, Vlad Ionescu), ThomasBS (Thomas Borregaard Sørensen, with KM [Kompetence Match]), pavankataria (Pavan Kataria), camwhite (Cameron White - YOSHICREW | ng2/node + mixing fire, Yoshicrew.com-out-loud contribution...), drmjg, 7ru57, abbyyyyy, aleksy, alexcode, andrevdb, anita, austingirl, ayvatan, baronliao, bazz3l, b3n0it, beee, benjones, benn, bigboom, bill_gates, blackburn1911, blackhatisme, bmw, boza, breitembach, ), cannap, chrifrachid, clint0n, code4life, codeotea, collins_badmus, cookiesliyr, chubz, crisey, cubor, daniel_foss, danielv, daratho, darkevil_s, david_joffe, denvercowboy, dgazail, diagram, didaum, dimer, dodakt, dopedopedope, drdaeman, dzetko, dzhekson6000, elina, euphemisminc, fauker, felix22, ffabiorj, flowdeep, flythiago, fork, frankly444, georgeb, gillie, gzj71ck4elm2, hailviral1 (Michael Lynn), hasanakyildiz, heuy_freeman, hsisbetterthanlol, hudwan, iamclaytonray, igsky, intelijens (Michael Bui), irondragon000, ishaan007lucky, jacobs, jaitaiwan, jamesxism, jarzi, jaycee, jenny0030, jeraqiv, jessy, jilizart, johnyblaze, jookeer, jpain, jrtruji, kaeris, karenkododyan, karrmarr, kaznowski, kimliksiz, kissloryshy, koen, kozmo, kranalex, LaffayetteBrittoPhoto (streamer with its C300 photo), lagroid, larrybrook, leonardosc, livecodingtv_support, livetester, lomoo (Matti [Matthias]), lukey_reborn78, magrib, marianyc, marko10, matroskinb, michaeljcalkins, michaelphp, milurabbit, mintisbad, monicaa, newton1, nuke, osa3, paezrice, peter_cados, patrick1, Phenotype3d, phprejext, pudottapommin, purpleizlegitx, raymond2, req1, rex2, rheghal, richard2, rooney2, rukewetony, samuraijackdavidson, sanyakozmey, sdwdewwe, silvaesouza, simplyronan, siox, skycis, skyfoxx, smart6307 (thunder will/canal2048), soulwish, sova, specter, stefanciprian, surebrec, systemis, szy, tarikogut, teras, thiagovgamg, tonycode, trello, usandy, v0n3x, Veiguiz (Daniel Veiguiz), wanket, wapaca, web-dev, web2ls, wilson2000, yanaseattle, zaaz, zakaria98, zann, zarytskiy, zeljko16 and zzzel (Marko Zel)  contribution!! jiangyubo530, kozmo, contribution!! Twitch contribution!! KindaFunnyGames (Greg, Colin, Tim, Nick and Kevin contribution..), AlfredoPlays (and his mates and Fwends playing Siege: Above_Perfection, koopz, Macky43, Hanson,  ... ), SoulFoam, Tectuma (GothicatCat1(Stacie), Tectuma(Kevin), Matty and Dylan), r00tkiller (Alexandre), 2016pocket, 4737, Alex_sn69, Alexetmartin, BaineGames, BarKatWhat, Barrydoodles, Bialyorzeu, cary205, cpmgamingltd, Curseappbot, Daratho, Darksoghatert, David_joffe, esmc103, Fai0thx, Fatality321, FullNerd33, GameDevCompany (Björn Swedish-Speaking-Finlandian), Gluttonys, Gmthiago, GrubbyJuicde, Handmadehero (Casey Muratori), Happyh0urr, HardlyDifficult, jhovgaard, Karaso00, kimliksiz, Kissloryshy, Ksifomaxos, Kycklingar, Letorusii, lion34590, Mattyplays6, MichaelJCalkins, Miga_, MOZG888VADIM, Niabi226, Ninebites, noochoom, nothings2 (Sean Barrett), Ojunbamo, Oryzapao, pixel_maniacs, Prafax, Pukateiubeste, Pyraxx, Roo220, seekaplayer, Shathan, Skintrader_pl Snipe0803, sundz98 (from Sweden), talus7, theD4T4, tophergates, Toshmnio, Tragonni, Vjzelynxykat, Wheeze202 (and hosting grumbul (British accent?)), XanderGamingTwich (Xander Jenkins), Zapanax and ZhootSe contribution!! Beam.pro contribution!! Igor_theBOSS196 contribution!! Slack contribution!! Zero Squad channel, Cameron White, trump, unicorn contribution!! h1dd3n, Remanos, RoxenRox, SeveredSnake, sharkbound and Xellar contribution!! bisba contribution!! Carol, Lúcio and Lucinho contribution!!
bilbo at veriverse
i7 6700 32GB RAM Windows 10 Pro
(booting also macOS Sierra, Hyper-V and Ubuntu 16.04 Unity Numix Theme Numix Circle Icons)

+++ appseed/aura/graphics/draw2d/draw2d_graphics.cpp
@@ -4108,7 +4108,7 @@ namespace draw2d
             while (i > 0)
             {
                sz = pgraphics->GetTextExtent(str, (int32_t)i);
-               if (sz.cx > rectClip.width())
+               if ((int) sz.cx > rectClip.width())
                {
                   i = ::str::utf8_dec(str, &((const char *)str)[i]) - ((const char *)str);
                   if (i <= 0)
